dvrp: extract DriveTask interface
