+++ contribs/drt/src/main/java/org/matsim/contrib/drt/schedule/DrtDriveTask.java
@@ -22,12 +22,12 @@ package org.matsim.contrib.drt.schedule;
 import static org.matsim.contrib.drt.schedule.DrtTaskBaseType.DRIVE;
 
 import org.matsim.contrib.dvrp.path.VrpPathWithTravelData;
-import org.matsim.contrib.dvrp.schedule.DriveTask;
+import org.matsim.contrib.dvrp.schedule.DefaultDriveTask;
 
 /**
  * @author michalm
  */
-public class DrtDriveTask extends DriveTask {
+public class DrtDriveTask extends DefaultDriveTask {
 	public static final DrtTaskType TYPE = new DrtTaskType(DRIVE);
 
 	public DrtDriveTask(VrpPathWithTravelData path, DrtTaskType taskType) {
