+++ contribs/dvrp/src/main/java/org/matsim/contrib/dvrp/schedule/DefaultDriveTask.java
@@ -17,21 +17,43 @@
  *                                                                         *
  * *********************************************************************** */
 
-package org.matsim.contrib.taxi.schedule;
+package org.matsim.contrib.dvrp.schedule;
 
-import static org.matsim.contrib.taxi.schedule.TaxiTaskBaseType.OCCUPIED_DRIVE;
-
+import org.matsim.contrib.dvrp.path.DivertedVrpPath;
+import org.matsim.contrib.dvrp.path.VrpPath;
 import org.matsim.contrib.dvrp.path.VrpPathWithTravelData;
-import org.matsim.contrib.dvrp.schedule.DefaultDriveTask;
-import org.matsim.contrib.taxi.passenger.TaxiRequest;
 
-public class TaxiOccupiedDriveTask extends DefaultDriveTask {
-	public static final TaxiTaskType TYPE = new TaxiTaskType(OCCUPIED_DRIVE);
+import com.google.common.base.MoreObjects;
 
-	public TaxiOccupiedDriveTask(VrpPathWithTravelData path, TaxiRequest request) {
-		super(TYPE, path);
-		if (request.getFromLink() != path.getFromLink() && request.getToLink() != path.getToLink()) {
-			throw new IllegalArgumentException();
+/**
+ * @author Michal Maciejewski (michalm)
+ */
+public class DefaultDriveTask extends AbstractTask implements DriveTask {
+	private VrpPath path;
+
+	public DefaultDriveTask(TaskType taskType, VrpPathWithTravelData path) {
+		super(taskType, path.getDepartureTime(), path.getArrivalTime());
+		this.path = path;
+	}
+
+	@Override
+	public final VrpPath getPath() {
+		return path;
+	}
+
+	@Override
+	public final void pathDiverted(DivertedVrpPath divertedPath, double newEndTime) {
+		// can only divert an ongoing task
+		if (getStatus() != TaskStatus.STARTED) {
+			throw new IllegalStateException();
 		}
+
+		path = divertedPath;
+		setEndTime(newEndTime);
+	}
+
+	@Override
+	public String toString() {
+		return MoreObjects.toStringHelper(this).add("super", super.toString()).add("path", path).toString();
 	}
 }
